refactor: Extract OtpTextField to its own file
